// app/Models/Employee.ts
export interface Employee {
    employeeName: string;
    employeeSurname: string;
    orgEmail: string;
    // orgName: string;
    orgTelephone: string; 
    personId: number;
    role: string;
    userTelephone?: string; 
    userId: number;
    username: string;
    orgId?: number;     
    password?: string;   
    telephone? :number;
    isActive: boolean; 
    warehouse?: string;
    
}